v2.4.6.8 -> v2.4.6.9
  - Dan Quinlan: cramfs update
  - Ben Collins: IEEE 1394 update
  - David Miller: network update (pppoe, routing cache stats), sparc32 update
  - me: only dump core once per threaded app (first one wins)
  - me: use new completion handlers for block device requests (same race
  as with vfork, see -pre7)

+++ fs/affs/amigaffs.c
@@ -284,7 +284,7 @@ affs_remove_header(struct dentry *dentry)
 
 	pr_debug("AFFS: remove_header(key=%ld)\n", inode->i_ino);
 	retval = -EIO;
-	bh = affs_bread(sb, (u32)dentry->d_fsdata);
+	bh = affs_bread(sb, (u32)(long)dentry->d_fsdata);
 	if (!bh)
 		goto done;
